--- pkg/renderer/const.go
@@ -1,6 +1,0 @@
-package renderer
-
-const (
-	contentTypeKey   = "Content-Type"
-	contentTypeValue = "application/json; charset=utf-8"
-)
